Use legal names instead of preferred names
There are a lot of students that have empty preferred names, so this isn't a reliable field to use here unless it can fall back to the legal name if empty. I'm not sure how to do that right now, or if its possible, so this is a good enough solution.