"Simplify" SQL query for students for IE Edu Forms (#19)
Despite this query being fairly complicated, this works around resource limitations on the MSS SQL server. Putting this all in a single query causes significant resource strain on the server, because we're already dealing with complex queries presented as views before we even start, and the person and address tables are likely some of the biggest tables in the system.